# Add adapter to improve Passenger performance