--- a/lib/faye/adapters/passenger.rb
+++ b/lib/faye/adapters/passenger.rb
@@ -1,0 +1,3 @@
+if defined?(PhusionPassenger)
+  PhusionPassenger.advertised_concurrency_level = 0
+end