Hi, 

if you don't override Passenger's advertised_concurrency value to be infinity, it will limit the amount of sessions per process to 1. This forces users to have a process per websocket connection (expensive) or to buy Phusion Passenger Enterprise and use a thread per websocket connection (slightly less expensive), both options bypass the cool performance boost EM based solutions should normally have.

When the app advertises infinite concurrency, you get all the benefits and can easily do thousands of websocket connections to a single process. I verified these patches by running npm's thor websocket stresser against the `passenger -start -R examples/config.ru`. 

Without adapter:
```
Durations (ms):

                     min     mean     stddev  median max    
Handshaking          9       177          87     179 328    
Latency              0       2             3       1 37     

Percentile (ms):

                      50%     66%     75%     80%     90%     95%     98%     98%    100%   
Handshaking          179     230     249     261     280     308     321     328     328    
Latency              1       2       3       3       5       6       10      12      37 
```

With adapter:
```
Durations (ms):

                     min     mean     stddev  median max    
Handshaking          10      41           19      41 91     
Latency              0       23           15      20 85     

Percentile (ms):

                      50%     66%     75%     80%     90%     95%     98%     98%    100%   
Handshaking          41      55      58      60      64      71      79      91      91     
Latency              20      24      31      33      44      53      72      77      85
```

The benchmarker isn't really well suited for benchmarking concurrency unfortunately, but you can see in the long max-duration that some sessions were held up for almost as long as the entire test ran. This doesn't happen with the concurrency enabled.