@d-snp 2015-02-23
The increased latencies are explained by the fact that as the sessions are now handled concurrently instead of sequentially the server is now busier whilst handling them.