@jcoglan 2015-02-23
Thank you for spotting this. I don't think it should be an 'adapter' though.

The adapters serve a very specific purpose: some app servers require patching in order to allow direct access to the TCP stream, which is necessary for implementing WebSocket. These patches must be loaded before any requests are processed, in order for requests to be correctly handled. That is, by the time you call `Faye::WebSocket.new` it's too late for the patch to be applied since the app server has already begun handling the request itself. Without these patches, `Faye::WebSocket` simply won't work.

This adapter doesn't fit into that category of 'ugly but necessary' patches. It's really a set of configuration, via the server's public API; an API that a user of this library is expected to interact with themselves, as opposed to the internal APIs that the other adapters patch.

I think libraries should not get involved in changing global settings if they can possibly avoid it, especially indirectly. What if the user has set this value themselves before loading `faye/websocket`?

Someone using Passenger is presumably using directly, rather than as a transitive dependency, and they should control its configuration rather than other libraries implicitly modifying it. We should by all means document the optimal settings for the app servers we support, but not enforce those ourselves.

I'm not sure if I've explained myself clearly, does this make sense?