@d-snp 2015-02-23
Hi, thanks you're totally making sense. I agree with you that semantically it's not really an 'adapter'. But the 'ugly but necessary' thing is the reason I thought this might apply anyway. The author of Passenger (my boss, Hongli) considers this configuration to be a hack and he doesn't want to publically document it (https://github.com/phusion/passenger/issues/1187). 

Hongli doesn't think it's a big deal that users have to make a configuration to work around some application server peculiarity, but I've recently started my own ruby websocket library (celluloid-websocket, using your driver), and I'm sure you agree that it totally sucks when your nice API example in the readme.md has to include these caveats for specific app servers. You worked around it nicely with the adapter api that shows the user there's some custom stuff happening for each app server, but at least they're not bothered with the details.

I'm really hoping I could get it into your project, this way or perhaps somewhere you think is more appropriate. If you really think we shouldn't be doing that I understand and I'll work on stressing the importance of the configuration in Passenger's websocket examples.