@jcoglan 2015-03-03
Okay, I take your point. However since it's an API addition it would need a minor version bump, which ordinarily wouldn't be a big deal, only the versioning of this module is in step with https://github.com/faye/faye-websocket-node -- I releases these on the same schedule and maintain feature parity. Bumping the minor version without any other new features that are ported to Node would break that scheme.

I'll pull this into a branch for now and document the Passenger config on `master` for the time being.